# Fix max_calls, streamline logic