--- a/tabrepo/nips2025_utils/artifacts/method_metadata.py
+++ b/tabrepo/nips2025_utils/artifacts/method_metadata.py
@@ -37,7 +37,6 @@ class MethodMetadata:
         has_processed: bool = False,
         has_results: bool = False,
         upload_as_public: bool = False,
-        custom_artifact_name: bool = False,
     ):
         self.method = method
         if artifact_name is None:
@@ -57,7 +56,6 @@ class MethodMetadata:
         if can_hpo is None:
             can_hpo = self.method_type == "config"
         self.can_hpo = can_hpo
-        self.custom_artifact_name = custom_artifact_name
 
         assert isinstance(self.method, str) and len(self.method) > 0
         assert isinstance(self.artifact_name, str) and len(self.artifact_name) > 0