--- a/tabrepo/nips2025_utils/end_to_end.py
+++ b/tabrepo/nips2025_utils/end_to_end.py
@@ -1,11 +1,9 @@
 from __future__ import annotations
 
-import os
 from pathlib import Path
 from typing_extensions import Self
 
 import pandas as pd
-from autogluon.common.savers import save_pd
 
 from tabrepo.benchmark.result import BaselineResult
 from tabrepo.nips2025_utils.artifacts.method_metadata import MethodMetadata
@@ -14,16 +12,11 @@ from tabrepo.nips2025_utils.end_to_end_single import (
     EndToEndResultsSingle,
     EndToEndSingle,
 )
-from tabrepo.nips2025_utils.fetch_metadata import load_task_metadata
 from tabrepo.nips2025_utils.method_processor import (
     generate_task_metadata,
     get_info_from_result,
-    load_all_artifacts,
     load_raw,
 )
-from tabrepo.nips2025_utils.tabarena_context import TabArenaContext
-from tabrepo.utils.pickle_utils import fetch_all_pickles_fast
-from tabrepo.utils.ray_utils import ray_map_list
 
 
 class EndToEnd:
@@ -141,111 +134,6 @@ class EndToEnd:
                 end_to_end.to_results() for end_to_end in self.end_to_end_lst
             ],
         )
-
-    @staticmethod
-    def create_and_cache_end_to_end_results(
-        path_raw: str | Path,
-        num_cpus: int | None = None,
-        artifact_name: str | None = None,
-    ) -> None:
-        """Create and cache end-to-end results for all methods in the given directory.
-
-        Args:
-            path_raw (str | Path): Path to the directory containing raw results.
-            num_cpus (int | None): Number of CPUs to use for parallel processing.
-                If None, it will use all available CPUs.
-            artifact_name (str | None): Optional name to distinguish different runs of
-                the same method.
-        """
-        if num_cpus is None:
-            num_cpus = len(os.sched_getaffinity(0))
-
-        print("Get results paths...")
-        all_file_paths_method = fetch_all_pickles_fast(
-            dir_path=path_raw, suffix="results.pkl"
-        )
-        print("Get task metadata...")
-        task_metadata = load_task_metadata()
-        # Below is too slow to use by default, TODO: get logic for any task that is fast
-        # task_metadata = generate_task_metadata(tids=list({r.split("/")[0] for r in all_file_paths_method}))
-
-        results = ray_map_list(
-            list_to_map=list(all_file_paths_method.values()),
-            func=_process_result_list,
-            func_element_key_string="file_paths_method",
-            num_workers=num_cpus,
-            num_cpus_per_worker=1,
-            func_put_kwargs={
-                "task_metadata": task_metadata,
-                "artifact_name": artifact_name,
-            },
-            track_progress=True,
-            tqdm_kwargs={"desc": "Processing Results"},
-            ray_remote_kwargs={"max_calls": 10},
-        )
-
-        print("Merging results...")
-        method_metadata, hpo_results, model_results = results[0]
-        for results_method in results[1:]:
-            method_metadata_other, hpo_results_other, model_results_other = (
-                results_method
-            )
-
-            # Capture the any() in metadata creation.
-            if method_metadata.is_bag or method_metadata_other.is_bag:
-                method_metadata.is_bag = True
-                method_metadata_other.is_bag = True
-
-            if method_metadata.__dict__ != method_metadata_other.__dict__:
-                raise ValueError(
-                    "Method metadata mismatch! "
-                    f"{method_metadata.__dict__} != {method_metadata_other.__dict__}"
-                )
-
-            # merge results
-            hpo_results = pd.concat([hpo_results, hpo_results_other], ignore_index=True)
-            model_results = pd.concat(
-                [model_results, model_results_other], ignore_index=True
-            )
-
-        print(f"Save metadata and results to {method_metadata.path}...")
-        method_metadata.to_yaml()
-        save_pd.save(path=str(method_metadata.path_results_hpo()), df=hpo_results)
-        save_pd.save(path=str(method_metadata.path_results_model()), df=model_results)
-
-
-def _process_result_list(
-    *,
-    file_paths_method: list[Path],
-    task_metadata: pd.DataFrame,
-    artifact_name: str | None,
-) -> tuple[MethodMetadata, pd.DataFrame, pd.DataFrame]:
-    results_lst = load_all_artifacts(
-        file_paths=file_paths_method, engine="sequential", progress_bar=False
-    )
-    # Get metadata
-    method_metadata: MethodMetadata = MethodMetadata.from_raw(
-        results_lst=results_lst, artifact_name=artifact_name
-    )
-
-    # Get evaluation repository
-    repo = method_metadata.generate_repo(
-        results_lst=results_lst,
-        task_metadata=task_metadata,
-    )
-
-    # Getting Tabarena context
-    tabarena_context = TabArenaContext()
-    tabarena_context.backend = "native"
-    tabarena_context.engine = "sequential"
-    hpo_results, model_results = tabarena_context.simulate_repo(
-        repo=repo,
-        method=method_metadata,
-        use_rf_config_fallback=False,
-        cache=False,
-    )
-
-    return method_metadata, hpo_results, model_results
 
 
 class EndToEndResults: