--- a/tabrepo/nips2025_utils/end_to_end.py
+++ b/tabrepo/nips2025_utils/end_to_end.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
+import os
 from pathlib import Path
+from typing import Literal
 from typing_extensions import Self
 
 import pandas as pd
@@ -12,11 +14,15 @@ from tabrepo.nips2025_utils.end_to_end_single import (
     EndToEndResultsSingle,
     EndToEndSingle,
 )
+from tabrepo.nips2025_utils.fetch_metadata import load_task_metadata
 from tabrepo.nips2025_utils.method_processor import (
     generate_task_metadata,
     get_info_from_result,
+    load_all_artifacts,
     load_raw,
 )
+from tabrepo.utils.pickle_utils import fetch_all_pickles
+from tabrepo.utils.ray_utils import ray_map_list
 
 
 class EndToEnd:
@@ -48,6 +54,7 @@ class EndToEnd:
         name: str | None = None,
         name_suffix: str | None = None,
         artifact_name: str | None = None,
+        backend: Literal["ray", "native"] = "ray",
         verbose: bool = True,
     ) -> Self:
         log = print if verbose else (lambda *a, **k: None)
@@ -85,6 +92,7 @@ class EndToEnd:
                 name=name,
                 name_suffix=name_suffix,
                 artifact_name=artifact_name,
+                backend=backend,
                 verbose=verbose,
             )
             end_to_end_lst.append(cur_end_to_end)
@@ -93,16 +101,18 @@ class EndToEnd:
     @classmethod
     def from_path_raw(
         cls,
-        path_raw: str | Path,
+        path_raw: str | Path | list[str | Path],
         task_metadata: pd.DataFrame | None = None,
         cache: bool = True,
         cache_raw: bool = True,
         name: str = None,
         name_suffix: str = None,
         artifact_name: str | None = None,
+        backend: Literal["ray", "native"] = "ray",
         verbose: bool = True,
     ) -> Self:
-        results_lst: list[BaselineResult] = load_raw(path_raw=path_raw)
+        engine = "ray" if backend == "ray" else "sequential"
+        results_lst: list[BaselineResult] = load_raw(path_raw=path_raw, engine=engine)
         return cls.from_raw(
             results_lst=results_lst,
             task_metadata=task_metadata,
@@ -111,6 +121,7 @@ class EndToEnd:
             name=name,
             name_suffix=name_suffix,
             artifact_name=artifact_name,
+            backend=backend,
             verbose=verbose,
         )
 
@@ -127,6 +138,78 @@ class EndToEnd:
             )
             end_to_end_lst.append(end_to_end_single)
         return cls(end_to_end_lst=end_to_end_lst)
+
+    @staticmethod
+    def create_and_cache_end_to_end_results(
+            path_raw: str | Path | list[str | Path],
+            num_cpus: int | None = None,
+            artifact_name: str | None = None,
+            task_metadata: pd.DataFrame | None = None,
+    ) -> EndToEndResults:
+        """Create and cache end-to-end results for all methods in the given directory.
+
+        Args:
+            path_raw (str | Path | list[str | Path): Path to the directory containing raw results.
+            num_cpus (int | None): Number of CPUs to use for parallel processing.
+                If None, it will use all available CPUs.
+            artifact_name (str | None): Optional name to distinguish different runs of
+                the same method.
+        """
+        if num_cpus is None:
+            num_cpus = len(os.sched_getaffinity(0))
+
+        print("Get results paths...")
+        file_paths = fetch_all_pickles(
+            dir_path=path_raw, suffix="results.pkl"
+        )
+
+        all_file_paths_method = {}
+        for file_path in file_paths:
+            did_sid = f"{file_path.parts[-3]}/{file_path.parts[-2]}"
+            if did_sid not in all_file_paths_method:
+                all_file_paths_method[did_sid] = []
+            all_file_paths_method[did_sid].append(file_path)
+
+        if task_metadata is None:
+            print("Get task metadata...")
+            task_metadata = load_task_metadata()
+            # Below is too slow to use by default, TODO: get logic for any task that is fast
+            # task_metadata = generate_task_metadata(tids=list({r.split("/")[0] for r in all_file_paths_method}))
+
+        results: list[EndToEndResults] = ray_map_list(
+            list_to_map=list(all_file_paths_method.values()),
+            func=_process_result_list,
+            func_element_key_string="file_paths_method",
+            num_workers=num_cpus,
+            num_cpus_per_worker=1,
+            func_put_kwargs={
+                "task_metadata": task_metadata,
+                "artifact_name": artifact_name,
+            },
+            track_progress=True,
+            tqdm_kwargs={"desc": "Processing Results"},
+            ray_remote_kwargs={"max_calls": 0},
+        )
+        results: list[EndToEndResultsSingle] = [
+            e2e_single for e2e in results for e2e_single in e2e.end_to_end_results_lst
+        ]
+
+        print("Merging results...")
+        results_per_method = {}
+        for e2e_single in results:
+            method = e2e_single.method_metadata.method
+            if method not in results_per_method:
+                results_per_method[method] = []
+            results_per_method[method].append(e2e_single)
+        e2e_single_lst = []
+        for method, e2e_lst in results_per_method.items():
+            cur_e2e = EndToEndResultsSingle.concat(e2e_lst=e2e_lst)
+            e2e_single_lst.append(cur_e2e)
+        e2e_results = EndToEndResults(end_to_end_results_lst=e2e_single_lst)
+
+        print(f"Save metadata and results...")
+        e2e_results.cache()
+        return e2e_results
 
     def to_results(self) -> EndToEndResults:
         return EndToEndResults(
@@ -232,3 +315,31 @@ class EndToEndResults:
             )
             end_to_end_results_lst.append(end_to_end_results)
         return cls(end_to_end_results_lst=end_to_end_results_lst)
+
+    def cache(self):
+        for e2e_results_single in self.end_to_end_results_lst:
+            e2e_results_single.cache()
+
+
+def _process_result_list(
+    *,
+    file_paths_method: list[Path],
+    task_metadata: pd.DataFrame,
+    artifact_name: str | None,
+) -> EndToEndResults:
+    results_lst = load_all_artifacts(
+        file_paths=file_paths_method, engine="sequential", progress_bar=False
+    )
+
+    e2e = EndToEnd.from_raw(
+        results_lst=results_lst,
+        task_metadata=task_metadata,
+        artifact_name=artifact_name,
+        cache=False,
+        cache_raw=False,
+        backend="native",
+        verbose=False,
+    )
+    e2e_results = e2e.to_results()
+
+    return e2e_results