*Issue #, if available:*

#206 

*Description of changes:*

Runtimes based on 96 CPU cores from raw data of LightGBM_aio_0808.

- Fix max_calls, speeds up `create_and_cache_end_to_end_results` from 93s -> 56s
- Ensure we reload repo from cache if `cache=True` in `EndToEndSingle.from_raw`, speedup of HPO simulation from 1400s -> 30s.
- Update paper_runner and paper_runner_tabarena to handle edge-cases in `create_and_cache_end_to_end_results` due to only running a single task at a time in parallel.
- Streamline `_process_result_list` to call `EndToEndSingle.from_raw`

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.